fix unstable split check
